added a slider component and a “multiplier” function to the plugin

+++ negate4/plugin.h
@@ -18,7 +18,7 @@ __PYX_EXTERN_C DL_IMPORT(void) getParamConfig(struct ParamConfig *);
 __PYX_EXTERN_C DL_IMPORT(void) pluginFunction(float *, int, int);
 __PYX_EXTERN_C DL_IMPORT(int) pluginisready(void);
 __PYX_EXTERN_C DL_IMPORT(void) setIntParam(char *, int);
-__PYX_EXTERN_C DL_IMPORT(void) setFloatParam(char *, int);
+__PYX_EXTERN_C DL_IMPORT(void) setFloatParam(char *, float);
 
 #endif /* !__PYX_HAVE_API__negate4__plugin */
